building-segmentation-app/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
torch==2.1.0
torchvision==0.16.0
segment-anything==1.0
git+https://github.com/facebookresearch/segment-anything-2.git
python-dotenv==1.0.0
aiofiles==23.2.1 